fix: use a single client for upgrades
Without this change, the upgrade process creates a new client for every
reconciliation. This was a problem for two reasons. First, there was no
way to close the client connection once reconcilitation was done, so
connections were left open, and exhausted trustd's FD limit. This
has since been fixed in upstream Talos. The second issue is not really
an "issue", but more of an optimization. There really is no need to
create a client every time. We could create one client, that
automatically renew's it's certificate, and adjusts the targeted node by
setting it in the request's context. This is what this change does.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>