feat: remove the need for node annotations
This removes the dependency on node annotations to figure out if a node
upgrade is current in progress. The "in progress" data is currently
storred in the CRD's status, and as an annotation on the node. In the
case that a user needs to override the in progress status for whatever
reason, they currently need to modify the CRD and the node. To keep
things simple, we now only use the CRD for node's "in progress" status.

Signed-off-by: Andrew Rynhard <andrew@andrewrynhard.com>